improve Identifier and reorganize move core types
