+++ crates/iota-sdk-ffi/src/types/mod.rs
@@ -10,11 +10,10 @@ pub mod events;
 pub mod execution_status;
 pub mod gas;
 pub mod iota_names;
+pub mod move_core;
 pub mod move_package;
 pub mod object;
 pub mod signature;
-pub mod struct_tag;
 pub mod transaction;
-pub mod type_tag;
 pub mod validator;
 pub mod version;
