+++ crates/iota-sdk-ffi/src/types/move_core/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod identifier;
+mod struct_tag;
+mod type_tag;
+
+pub use identifier::Identifier;
+pub use struct_tag::StructTag;
+pub use type_tag::TypeTag;
