+++ crates/iota-sdk-ffi/src/types/move_core/struct_tag.rs
@@ -3,87 +3,10 @@
 
 use std::sync::Arc;
 
-use crate::{
-    error::Result,
-    types::{address::Address, type_tag::TypeTag},
+use crate::types::{
+    address::Address,
+    move_core::{identifier::Identifier, type_tag::TypeTag},
 };
-
-/// A move identifier
-///
-/// # BCS
-///
-/// The BCS serialized form for this type is defined by the following ABNF:
-///
-/// ```text
-/// identifier = %x01-80    ; length of the identifier
-///              (ALPHA *127(ALPHA / DIGIT / UNDERSCORE)) /
-///              (UNDERSCORE 1*127(ALPHA / DIGIT / UNDERSCORE))
-///
-/// UNDERSCORE = %x95
-/// ```
-#[derive(Debug, PartialEq, Eq, Hash, derive_more::From, derive_more::Display, uniffi::Object)]
-#[uniffi::export(Debug, Display, Eq, Hash)]
-pub struct Identifier(pub iota_sdk::types::Identifier);
-
-#[uniffi::export]
-impl Identifier {
-    #[uniffi::constructor]
-    pub fn new(identifier: String) -> Result<Self> {
-        Ok(Self(iota_sdk::types::Identifier::new(identifier)?))
-    }
-
-    pub fn as_str(&self) -> String {
-        self.0.as_str().to_owned()
-    }
-}
-
-macro_rules! export_struct_tag_ctors {
-    ($($name:ident),+ $(,)?) => { paste::paste! {
-        #[uniffi::export]
-        impl StructTag {$(
-            #[uniffi::constructor]
-            pub fn [< new_ $name:snake >]() -> Self {
-                Self(iota_sdk::types::StructTag::[< new_ $name:snake >]())
-            }
-
-            pub fn [< is_ $name:snake >](&self) -> bool {
-                self.0.[< is_ $name:snake >]()
-            }
-        )+}
-    } }
-}
-
-macro_rules! export_struct_tag_from_type_tag_ctors {
-    ($($name:ident),+ $(,)?) => { paste::paste! {
-        #[uniffi::export]
-        impl StructTag {$(
-            #[uniffi::constructor]
-            pub fn [< new_ $name:snake >](type_tag: &TypeTag) -> Self {
-                Self(iota_sdk::types::StructTag::[< new_ $name:snake >](type_tag.0.clone()))
-            }
-
-            pub fn [< is_ $name:snake >](&self) -> bool {
-                self.0.[< is_ $name:snake >]()
-            }
-        )+}
-    } }
-}
-
-macro_rules! export_struct_tag_from_struct_tag_ctors {
-    ($($name:ident),+ $(,)?) => { paste::paste! {
-        #[uniffi::export]
-        impl StructTag {$(
-            #[uniffi::constructor]
-            pub fn [< new_ $name:snake >](struct_tag: &StructTag) -> Self {
-                Self(iota_sdk::types::StructTag::[< new_ $name:snake >](struct_tag.0.clone()))
-            }
-
-            pub fn [< is_ $name:snake >](&self) -> bool {
-                self.0.[< is_ $name:snake >]()
-            }
-        )+}
-    } }
-}
 
 /// Type information for a move struct
 ///
@@ -188,6 +111,54 @@ impl StructTag {
     pub fn to_canonical_string(&self, with_prefix: bool) -> String {
         self.0.to_canonical_string(with_prefix)
     }
+}
+
+macro_rules! export_struct_tag_ctors {
+    ($($name:ident),+ $(,)?) => { paste::paste! {
+        #[uniffi::export]
+        impl StructTag {$(
+            #[uniffi::constructor]
+            pub fn [< new_ $name:snake >]() -> Self {
+                Self(iota_sdk::types::StructTag::[< new_ $name:snake >]())
+            }
+
+            pub fn [< is_ $name:snake >](&self) -> bool {
+                self.0.[< is_ $name:snake >]()
+            }
+        )+}
+    } }
+}
+
+macro_rules! export_struct_tag_from_type_tag_ctors {
+    ($($name:ident),+ $(,)?) => { paste::paste! {
+        #[uniffi::export]
+        impl StructTag {$(
+            #[uniffi::constructor]
+            pub fn [< new_ $name:snake >](type_tag: &TypeTag) -> Self {
+                Self(iota_sdk::types::StructTag::[< new_ $name:snake >](type_tag.0.clone()))
+            }
+
+            pub fn [< is_ $name:snake >](&self) -> bool {
+                self.0.[< is_ $name:snake >]()
+            }
+        )+}
+    } }
+}
+
+macro_rules! export_struct_tag_from_struct_tag_ctors {
+    ($($name:ident),+ $(,)?) => { paste::paste! {
+        #[uniffi::export]
+        impl StructTag {$(
+            #[uniffi::constructor]
+            pub fn [< new_ $name:snake >](struct_tag: &StructTag) -> Self {
+                Self(iota_sdk::types::StructTag::[< new_ $name:snake >](struct_tag.0.clone()))
+            }
+
+            pub fn [< is_ $name:snake >](&self) -> bool {
+                self.0.[< is_ $name:snake >]()
+            }
+        )+}
+    } }
 }
 
 export_struct_tag_ctors!(
