+++ crates/iota-sdk-ffi/src/types/transaction/mod.rs
@@ -13,11 +13,10 @@ use crate::{
         crypto::Bls12381PublicKey,
         digest::Digest,
         events::Event,
+        move_core::{Identifier, TypeTag},
         object::{GenesisObject, ObjectId, ObjectReference},
         signature::UserSignature,
-        struct_tag::Identifier,
         transaction::v1::TransactionEffectsV1,
-        type_tag::TypeTag,
         version::Version,
     },
 };
