+++ crates/iota-sdk-types/src/iota_names/mod.rs
@@ -10,7 +10,7 @@ pub mod registry;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 pub use self::name::{Name, NameFormat};
-use crate::{Address, ObjectId, StructTag, type_tag::IdentifierRef};
+use crate::{Address, ObjectId, StructTag, move_core::Identifier};
 
 /// An object to manage a second-level name (SLN).
 #[derive(Debug, Clone, Eq, PartialEq)]
@@ -57,8 +57,8 @@ impl SubnameRegistration {
 
 /// Unifying trait for [`NameRegistration`] and [`SubnameRegistration`]
 pub trait IotaNamesNft {
-    const MODULE: &IdentifierRef;
-    const TYPE_NAME: &IdentifierRef;
+    const MODULE: Identifier;
+    const TYPE_NAME: Identifier;
 
     fn type_(package_id: Address) -> StructTag {
         StructTag::new(package_id, Self::MODULE, Self::TYPE_NAME, Vec::new())
@@ -82,8 +82,8 @@ pub trait IotaNamesNft {
 }
 
 impl IotaNamesNft for NameRegistration {
-    const MODULE: &IdentifierRef = IdentifierRef::const_new("name_registration");
-    const TYPE_NAME: &IdentifierRef = IdentifierRef::const_new("NameRegistration");
+    const MODULE: Identifier = Identifier::from_static("name_registration");
+    const TYPE_NAME: Identifier = Identifier::from_static("NameRegistration");
 
     fn name(&self) -> &Name {
         &self.name
@@ -103,8 +103,8 @@ impl IotaNamesNft for NameRegistration {
 }
 
 impl IotaNamesNft for SubnameRegistration {
-    const MODULE: &IdentifierRef = IdentifierRef::const_new("subname_registration");
-    const TYPE_NAME: &IdentifierRef = IdentifierRef::const_new("SubnameRegistration");
+    const MODULE: Identifier = Identifier::from_static("subname_registration");
+    const TYPE_NAME: Identifier = Identifier::from_static("SubnameRegistration");
 
     fn name(&self) -> &Name {
         self.nft.name()
