+++ crates/iota-sdk-types/src/move_core/mod.rs
@@ -1,0 +1,29 @@
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod parse;
+
+mod identifier;
+#[cfg(feature = "serde")]
+#[cfg_attr(doc_cfg, doc(cfg(feature = "serde")))]
+mod serialization;
+mod struct_tag;
+mod type_tag;
+
+pub use identifier::Identifier;
+pub use struct_tag::StructTag;
+pub use type_tag::TypeTag;
+
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub struct TypeParseError {
+    pub source: String,
+}
+
+impl std::fmt::Display for TypeParseError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::fmt::Debug::fmt(self, f)
+    }
+}
+
+impl std::error::Error for TypeParseError {}
