+++ crates/iota-sdk-types/src/move_core/struct_tag.rs
@@ -1,0 +1,431 @@
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::{Address, Identifier, TypeParseError, TypeTag};
+
+/// Type information for a move struct
+///
+/// # BCS
+///
+/// The BCS serialized form for this type is defined by the following ABNF:
+///
+/// ```text
+/// struct-tag = address            ; address of the package
+///              identifier         ; name of the module
+///              identifier         ; name of the type
+///              (vector type-tag)  ; type parameters
+/// ```
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "proptest", derive(test_strategy::Arbitrary))]
+pub struct StructTag {
+    pub(crate) address: Address,
+    pub(crate) module: Identifier,
+    pub(crate) name: Identifier,
+    #[cfg_attr(feature = "proptest", strategy(proptest::strategy::Just(Vec::new())))]
+    pub(crate) type_params: Vec<TypeTag>,
+}
+
+impl StructTag {
+    pub fn new(
+        address: impl Into<Address>,
+        module: impl Into<Identifier>,
+        name: impl Into<Identifier>,
+        type_params: Vec<TypeTag>,
+    ) -> Self {
+        Self {
+            address: address.into(),
+            module: module.into(),
+            name: name.into(),
+            type_params,
+        }
+    }
+
+    /// Returns whether the type (excluding the type params) is the same as
+    /// another struct tag.
+    pub fn is_same_type_as(&self, other: &StructTag) -> bool {
+        self.address == other.address && self.module == other.module && self.name == other.name
+    }
+
+    /// Creates a new framework IOTA coin struct tag (`0x2::iota::IOTA`)
+    pub fn new_iota_coin_type() -> Self {
+        Self {
+            address: Address::FRAMEWORK,
+            module: Identifier::from_static("iota").into(),
+            name: Identifier::from_static("IOTA").into(),
+            type_params: vec![],
+        }
+    }
+
+    /// Checks if this is a framework IOTA coin type (`0x2::iota::IOTA`)
+    pub fn is_iota_coin_type(&self) -> bool {
+        self.address == Address::FRAMEWORK
+            && self.module == Identifier::from_static("iota")
+            && self.name == Identifier::from_static("IOTA")
+            && self.type_params.is_empty()
+    }
+
+    /// Creates a new framework gas coin struct tag
+    /// (`0x2::coin::Coin<0x2::iota::IOTA>`)
+    pub fn new_gas_coin() -> Self {
+        Self::new_coin(Self::new_iota_coin_type())
+    }
+
+    /// Creates a new framework coin struct tag (`0x2::coin::Coin<CoinType>`)
+    pub fn is_gas_coin(&self) -> bool {
+        self.address == Address::FRAMEWORK
+            && self.module == Identifier::from_static("coin")
+            && self.name == Identifier::from_static("Coin")
+            && matches!(
+                self.type_params.first(),
+                Some(TypeTag::Struct(boxed_struct_tag))
+                if boxed_struct_tag.is_iota_coin_type()
+            )
+    }
+
+    /// Creates a new framework object ID struct tag (`0x2::object::ID`)
+    pub fn new_id() -> Self {
+        Self {
+            address: Address::FRAMEWORK,
+            module: Identifier::from_static("object").into(),
+            name: Identifier::from_static("ID").into(),
+            type_params: vec![],
+        }
+    }
+
+    /// Checks if this is a framework object ID type (`0x2::object::ID`)
+    pub fn is_id(&self) -> bool {
+        self.address == Address::FRAMEWORK
+            && self.module == Identifier::from_static("object")
+            && self.name == Identifier::from_static("ID")
+    }
+
+    /// Creates a new framework object UID struct tag (`0x2::object::UID`)
+    pub fn new_uid() -> Self {
+        Self {
+            address: Address::FRAMEWORK,
+            module: Identifier::from_static("object").into(),
+            name: Identifier::from_static("UID").into(),
+            type_params: vec![],
+        }
+    }
+
+    /// Checks if this is a framework object UID type (`0x2::object::UID`)
+    pub fn is_uid(&self) -> bool {
+        self.address == Address::FRAMEWORK
+            && self.module == Identifier::from_static("object")
+            && self.name == Identifier::from_static("UID")
+    }
+
+    pub fn new_name(address: Address) -> Self {
+        Self {
+            address,
+            module: Identifier::from_static("name").into(),
+            name: Identifier::from_static("Name").into(),
+            type_params: vec![],
+        }
+    }
+
+    /// Creates a new dynamic field struct tag
+    /// (`0x2::dynamic_field::Field<KeyType, ValueType>`)
+    pub fn new_dynamic_field(key: impl Into<TypeTag>, value: impl Into<TypeTag>) -> Self {
+        Self {
+            address: Address::FRAMEWORK,
+            module: Identifier::from_static("dynamic_field").into(),
+            name: Identifier::from_static("Field").into(),
+            type_params: vec![key.into(), value.into()],
+        }
+    }
+
+    /// Checks if this is a Dynamic Field type
+    /// (`0x2::dynamic_field::Field<KeyType, ValueType>`)
+    pub fn is_dynamic_field(&self) -> bool {
+        self.address == Address::FRAMEWORK
+            && self.module == Identifier::from_static("dynamic_field")
+            && self.name == Identifier::from_static("Field")
+    }
+
+    /// Returns the coin type if this is a Coin type, None otherwise
+    pub fn coin_type_opt(&self) -> Option<&crate::TypeTag> {
+        let Self {
+            address,
+            module,
+            name,
+            type_params,
+        } = self;
+
+        if address == &Address::FRAMEWORK
+            && module == "coin"
+            && name == "Coin"
+            && type_params.len() == 1
+        {
+            type_params.first()
+        } else {
+            None
+        }
+    }
+
+    /// Returns the coin type of this `StructTag`, panics if not a coin
+    pub fn coin_type(&self) -> &TypeTag {
+        self.coin_type_opt().expect("not a coin")
+    }
+
+    /// Returns the address part of a `StructTag`
+    pub fn address(&self) -> Address {
+        self.address
+    }
+
+    /// Returns the module part of a `StructTag`
+    pub fn module(&self) -> &Identifier {
+        &self.module
+    }
+
+    /// Returns the name part of a `StructTag`
+    pub fn name(&self) -> &Identifier {
+        &self.name
+    }
+
+    /// Returns the type params part of a `StructTag`
+    pub fn type_params(&self) -> &[TypeTag] {
+        &self.type_params
+    }
+
+    /// Decomposes the StructTag into its parts
+    pub fn into_parts(self) -> (Address, Identifier, Identifier, Vec<TypeTag>) {
+        (self.address, self.module, self.name, self.type_params)
+    }
+
+    /// Returns the string representation of this struct tag using the
+    /// canonical display, with or without a `0x` prefix.
+    pub fn to_canonical_string(&self, with_prefix: bool) -> String {
+        let mut tag = format!(
+            "{}::{}::{}",
+            self.address.to_canonical_string(with_prefix),
+            self.module,
+            self.name
+        );
+        if let Some(first_type) = self.type_params.first() {
+            tag.push_str(&format!("<{first_type}"));
+            for ty in self.type_params.iter().skip(1) {
+                tag.push_str(&format!(", {ty}"));
+            }
+            tag.push('>');
+        }
+        tag
+    }
+}
+
+macro_rules! add_struct_tag_ctor {
+    (@with_module $address:ident, $($module:ident :: $name:ident),+ $(,)?) => {
+        $(
+            paste::paste! {
+                pub fn [< new_ $module:snake _ $name:snake >]() -> Self {
+                    Self {
+                        address: Address::$address,
+                        module: Identifier::from_static(stringify!($module)).into(),
+                        name: Identifier::from_static(stringify!($name)).into(),
+                        type_params: vec![],
+                    }
+                }
+
+                pub fn [< is_ $module:snake _ $name:snake >](&self) -> bool {
+                    self.address == Address::$address
+                        && self.module == Identifier::from_static(stringify!($module))
+                        && self.name == Identifier::from_static(stringify!($name))
+                }
+            }
+        )+
+    };
+    ($address:ident, $($module:ident :: $name:ident),+ $(,)?) => {
+        $(
+            paste::paste! {
+                pub fn [< new_ $name:snake >]() -> Self {
+                    Self {
+                        address: Address::$address,
+                        module: Identifier::from_static(stringify!($module)).into(),
+                        name: Identifier::from_static(stringify!($name)).into(),
+                        type_params: vec![],
+                    }
+                }
+
+                pub fn [< is_ $name:snake >](&self) -> bool {
+                    self.address == Address::$address
+                        && self.module == Identifier::from_static(stringify!($module))
+                        && self.name == Identifier::from_static(stringify!($name))
+                }
+            }
+        )+
+    };
+}
+
+macro_rules! add_struct_tag_ctor_from_struct_tag {
+    (@with_module $address:ident, $($module:ident :: $name:ident),+ $(,)?) => {
+        $(
+            paste::paste! {
+                pub fn [< new_ $module:snake _ $name:snake >](struct_tag: impl Into<StructTag>) -> Self {
+                    Self {
+                        address: Address::$address,
+                        module: Identifier::from_static(stringify!($module)).into(),
+                        name: Identifier::from_static(stringify!($name)).into(),
+                        type_params: vec![TypeTag::Struct(Box::new(struct_tag.into()))],
+                    }
+                }
+
+                pub fn [< is_ $module:snake _ $name:snake >](&self) -> bool {
+                    self.address == Address::$address
+                        && self.module == Identifier::from_static(stringify!($module))
+                        && self.name == Identifier::from_static(stringify!($name))
+                        && self.type_params.len() == 1
+                }
+            }
+        )+
+    };
+    ($address:ident, $($module:ident :: $name:ident),+ $(,)?) => {
+        $(
+            paste::paste! {
+                pub fn [< new_ $name:snake >](struct_tag: impl Into<StructTag>) -> Self {
+                    Self {
+                        address: Address::$address,
+                        module: Identifier::from_static(stringify!($module)).into(),
+                        name: Identifier::from_static(stringify!($name)).into(),
+                        type_params: vec![TypeTag::Struct(Box::new(struct_tag.into()))],
+                    }
+                }
+
+                pub fn [< is_ $name:snake >](&self) -> bool {
+                    self.address == Address::$address
+                        && self.module == Identifier::from_static(stringify!($module))
+                        && self.name == Identifier::from_static(stringify!($name))
+                        && self.type_params.len() == 1
+                }
+            }
+        )+
+    };
+}
+
+macro_rules! add_struct_tag_ctor_from_type_tag {
+    (@with_module $address:ident, $($module:ident :: $name:ident),+ $(,)?) => {
+        $(
+            paste::paste! {
+                pub fn [< new_ $module:snake _ $name:snake >](type_tag: impl Into<TypeTag>) -> Self {
+                    Self {
+                        address: Address::$address,
+                        module: Identifier::from_static(stringify!($module)).into(),
+                        name: Identifier::from_static(stringify!($name)).into(),
+                        type_params: vec![type_tag.into()],
+                    }
+                }
+
+                pub fn [< is_ $module:snake _ $name:snake >](&self) -> bool {
+                    self.address == Address::$address
+                        && self.module == Identifier::from_static(stringify!($module))
+                        && self.name == Identifier::from_static(stringify!($name))
+                        && self.type_params.len() == 1
+                }
+            }
+        )+
+    };
+    ($address:ident, $($module:ident :: $name:ident),+ $(,)?) => {
+        $(
+            paste::paste! {
+                pub fn [< new_ $name:snake >](type_tag: impl Into<TypeTag>) -> Self {
+                    Self {
+                        address: Address::$address,
+                        module: Identifier::from_static(stringify!($module)).into(),
+                        name: Identifier::from_static(stringify!($name)).into(),
+                        type_params: vec![type_tag.into()],
+                    }
+                }
+
+                pub fn [< is_ $name:snake >](&self) -> bool {
+                    self.address == Address::$address
+                        && self.module == Identifier::from_static(stringify!($module))
+                        && self.name == Identifier::from_static(stringify!($name))
+                        && self.type_params.len() == 1
+                }
+            }
+        )+
+    };
+}
+
+impl StructTag {
+    add_struct_tag_ctor!(
+        FRAMEWORK,
+        clock::Clock,
+        config::Config,
+        iota::IotaTreasuryCap,
+        package::UpgradeCap,
+        package::UpgradeTicket,
+        package::UpgradeReceipt,
+        system_admin_cap::IotaSystemAdminCap,
+        url::Url,
+        bag::Bag,
+        object_bag::ObjectBag,
+    );
+    add_struct_tag_ctor!(@with_module FRAMEWORK, deny_list::ConfigKey, deny_list::AddressKey, deny_list::GlobalPauseKey, transfer::Receiving);
+    add_struct_tag_ctor!(
+        SYSTEM,
+        iota_system::IotaSystemState,
+        staking_pool::StakedIota,
+        timelocked_staking::TimelockedStakedIota,
+        iota_system_state_inner::SystemEpochInfoEvent
+    );
+    add_struct_tag_ctor!(STD, string::String);
+    add_struct_tag_ctor!(@with_module STD, ascii::String);
+    add_struct_tag_ctor_from_struct_tag!(
+        FRAMEWORK,
+        coin::CoinMetadata,
+        coin::TreasuryCap,
+        coin_manager::CoinManager,
+        display::DisplayCreated
+    );
+    add_struct_tag_ctor_from_struct_tag!(@with_module FRAMEWORK, display::VersionUpdated);
+    add_struct_tag_ctor_from_type_tag!(FRAMEWORK, coin::Coin, balance::Balance, timelock::TimeLock);
+    add_struct_tag_ctor_from_type_tag!(@with_module FRAMEWORK, config::Setting, dynamic_object_field::Wrapper);
+    add_struct_tag_ctor_from_type_tag!(STD, option::Option);
+}
+
+impl std::fmt::Display for StructTag {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}::{}::{}",
+            self.address.to_short_hex(),
+            self.module,
+            self.name
+        )?;
+        if let Some(first_type) = self.type_params.first() {
+            write!(f, "<{first_type}")?;
+            for ty in self.type_params.iter().skip(1) {
+                write!(f, ", {ty}")?;
+            }
+            write!(f, ">")?;
+        }
+        Ok(())
+    }
+}
+
+impl std::str::FromStr for StructTag {
+    type Err = TypeParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        crate::move_core::parse::parse_struct_tag(s)
+            .map_err(|_| TypeParseError { source: s.into() })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_struct_tag_parsing() {
+        let tag_str = "0x2::coin::Coin<0x2::iota::IOTA>";
+        let struct_tag: StructTag = tag_str.parse().unwrap();
+        assert!(struct_tag.is_coin());
+        let coin_type = struct_tag.coin_type();
+        assert!(coin_type.is_struct());
+        let coin_struct = coin_type.as_struct_tag();
+        assert!(coin_struct.is_iota_coin_type());
+    }
+}
